init: add a simple container server
This is the first commit with code for a simple server structure based
on the "Docker and Kubernetes: The Complete Guide" course [0], by
Stephen Grider.

The code sets up a conteiner with a redis server image that will serve a
Node.js code running from another container.

This code was tested with Node version v10.10.0 and docker-compose
version 1.22.0, build f46880fe.

Ref.: Lecture 51

[0] https://www.udemy.com/share/100r4uAkMcdlZTTHw=/

Signed-off-by: Daniel Kreling <dkreling@daitangroup.com>